openhands-sdk/pyproject.toml
[project]
name = "openhands-sdk"
version = "1.16.1"
description = "OpenHands SDK - Core functionality for building AI agents"

requires-python = ">=3.12"
dependencies = [
    "agent-client-protocol>=0.8.1",
    "deprecation>=2.1.0",
    "fakeredis[lua]>=2.32.1",  # Explicit dependency for docket/fastmcp background tasks
    "fastmcp>=3.0.0",
    "filelock>=3.20.1",
    "httpx[socks]>=0.27.0",
    "litellm==1.80.10",
    "pydantic>=2.12.5",
    "python-frontmatter>=1.1.0",
    "python-json-logger>=3.3.0",
    "tenacity>=9.1.2",
    "websockets>=12",
    "lmnr>=0.7.24",
]

[project.urls]
Source = "https://github.com/OpenHands/software-agent-sdk"
Homepage = "https://github.com/OpenHands/software-agent-sdk"
Documentation = "https://docs.openhands.dev/sdk"
"Bug Tracker" = "https://github.com/OpenHands/software-agent-sdk/issues"

[project.optional-dependencies]
boto3 = ["boto3>=1.35.0"]
pillow = ["pillow>=12.1.1"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
include = ["openhands.sdk*"]
namespaces = true

[tool.setuptools.package-data]
"*" = ["py.typed", "*.j2"]
